Beats: Merge new beats implementation into abstract Beats class
