+++ src/track/beats.h
@@ -20,15 +20,39 @@
 namespace mixxx {
 
 class Beats;
+class BeatIterator;
 typedef std::shared_ptr<const Beats> BeatsPointer;
 
-class BeatIterator {
+class BeatMarker {
   public:
-    virtual ~BeatIterator() = default;
-    virtual bool hasNext() const = 0;
-    virtual audio::FramePos next() = 0;
+    BeatMarker(mixxx::audio::FramePos position, int beatsTillNextMarker)
+            : m_position(position), m_beatsTillNextMarker(beatsTillNextMarker) {
+        DEBUG_ASSERT(m_position.isValid());
+        DEBUG_ASSERT(!m_position.isFractional());
+        DEBUG_ASSERT(m_beatsTillNextMarker > 0);
+    }
+
+    mixxx::audio::FramePos position() const {
+        return m_position;
+    }
+
+    int beatsTillNextMarker() const {
+        return m_beatsTillNextMarker;
+    }
+
+  private:
+    mixxx::audio::FramePos m_position;
+    int m_beatsTillNextMarker;
 };
 
+inline bool operator==(const BeatMarker& lhs, const BeatMarker& rhs) {
+    return (lhs.position() == rhs.position() &&
+            lhs.beatsTillNextMarker() == rhs.beatsTillNextMarker());
+}
+
+inline bool operator!=(const BeatMarker& lhs, const BeatMarker& rhs) {
+    return !(lhs == rhs);
+}
 /// Beats is the base class for BPM and beat management classes. It provides a
 /// specification of all methods a beat-manager class must provide, as well as
 /// a capability model for representing optional features.
@@ -36,7 +60,148 @@ class BeatIterator {
 /// All instances of this class are supposed to be managed by std::shared_ptr!
 class Beats : private std::enable_shared_from_this<Beats> {
   public:
-    virtual ~Beats() = default;
+    class ConstIterator {
+      public:
+        using iterator_category = std::random_access_iterator_tag;
+        using value_type = mixxx::audio::FramePos;
+        using difference_type = int;
+        using pointer = value_type*;
+        using reference = value_type&;
+
+        ConstIterator(const Beats* beats,
+                std::vector<BeatMarker>::const_iterator it,
+                int beatOffset)
+                : m_beats(beats), m_it(it), m_beatOffset(beatOffset) {
+            updateValue();
+        }
+
+        mixxx::audio::FrameDiff_t beatLengthFrames() const;
+
+        // Iterator methods
+
+        const value_type& operator*() const {
+            return m_value;
+        }
+
+        const value_type* operator->() const {
+            return &m_value;
+        }
+
+        ConstIterator& operator++() {
+            *this += 1;
+            return *this;
+        }
+
+        ConstIterator& operator--() {
+            *this -= 1;
+            return *this;
+        }
+
+        ConstIterator operator++(difference_type) {
+            ConstIterator tmp = *this;
+            ++(*this);
+            return tmp;
+        }
+
+        ConstIterator operator--(difference_type) {
+            ConstIterator tmp = *this;
+            --(*this);
+            return tmp;
+        }
+
+        ConstIterator operator+(difference_type n) {
+            ConstIterator tmp = *this;
+            tmp += n;
+            return tmp;
+        }
+
+        ConstIterator operator-(difference_type n) {
+            ConstIterator tmp = *this;
+            tmp -= n;
+            return tmp;
+        }
+
+        ConstIterator operator+=(difference_type n);
+        ConstIterator operator-=(difference_type n);
+
+        difference_type operator-(const ConstIterator& other) const;
+
+        friend bool operator==(const ConstIterator& lhs, const ConstIterator& rhs) {
+            return lhs.m_beats == rhs.m_beats && lhs.m_it == rhs.m_it &&
+                    lhs.m_beatOffset == rhs.m_beatOffset;
+        }
+
+        friend bool operator!=(const ConstIterator& lhs, const ConstIterator& rhs) {
+            return !(lhs == rhs);
+        }
+
+      private:
+        void updateValue();
+
+        mixxx::audio::FramePos m_value;
+
+        const Beats* m_beats;
+        std::vector<BeatMarker>::const_iterator m_it;
+        int m_beatOffset;
+    };
+
+    Beats(std::vector<BeatMarker> markers,
+            mixxx::audio::FramePos endMarkerPosition,
+            mixxx::Bpm endMarkerBpm,
+            mixxx::audio::SampleRate sampleRate,
+            const QString& subVersion)
+            : m_markers(std::move(markers)),
+              m_endMarkerPosition(endMarkerPosition),
+              m_endMarkerBpm(endMarkerBpm),
+              m_sampleRate(sampleRate),
+              m_subVersion(subVersion) {
+        DEBUG_ASSERT(m_endMarkerPosition.isValid());
+        DEBUG_ASSERT(!m_endMarkerPosition.isFractional());
+        DEBUG_ASSERT(m_endMarkerBpm.isValid());
+        DEBUG_ASSERT(m_sampleRate.isValid());
+    }
+
+    Beats(mixxx::audio::FramePos endMarkerPosition,
+            mixxx::Bpm endMarkerBpm,
+            mixxx::audio::SampleRate sampleRate,
+            const QString& subVersion)
+            : Beats(std::vector<BeatMarker>(),
+                      endMarkerPosition,
+                      endMarkerBpm,
+                      sampleRate,
+                      subVersion) {
+    }
+
+    ~Beats() = default;
+
+    ConstIterator cbegin() const {
+        return ConstIterator(this, m_markers.cbegin(), 0);
+    }
+
+    ConstIterator cend() const {
+        return ConstIterator(this, m_markers.cend(), 1);
+    }
+
+    ConstIterator cearliest() const {
+        return ConstIterator(this, m_markers.cbegin(), std::numeric_limits<int>::lowest());
+    }
+
+    ConstIterator clatest() const {
+        return ConstIterator(this, m_markers.cend(), std::numeric_limits<int>::max());
+    }
+
+    ConstIterator iteratorFrom(audio::FramePos position) const;
+
+    friend bool operator==(const Beats& lhs, const Beats& rhs) {
+        return lhs.m_markers == rhs.m_markers &&
+                lhs.m_endMarkerPosition == rhs.m_endMarkerPosition &&
+                lhs.m_endMarkerBpm == rhs.m_endMarkerBpm && lhs.m_sampleRate &&
+                rhs.m_sampleRate;
+    }
+
+    friend bool operator!=(const Beats& lhs, const Beats& rhs) {
+        return !(lhs == rhs);
+    }
 
     BeatsPointer clonePointer() const {
         // All instances are immutable and can be shared safely
@@ -48,6 +213,16 @@ class Beats : private std::enable_shared_from_this<Beats> {
             const QString& beatsVersion,
             const QString& beatsSubVersion,
             const QByteArray& beatsSerialized);
+
+    static BeatsPointer fromBeatGridByteArray(
+            audio::SampleRate sampleRate,
+            const QString& subVersion,
+            const QByteArray& byteArray);
+
+    static BeatsPointer fromBeatMapByteArray(
+            audio::SampleRate sampleRate,
+            const QString& subVersion,
+            const QByteArray& byteArray);
 
     static mixxx::BeatsPointer fromConstTempo(
             audio::SampleRate sampleRate,
@@ -74,18 +249,22 @@ class Beats : private std::enable_shared_from_this<Beats> {
     /// TODO: This is only needed for the "Asumme Constant Tempo" checkbox in
     /// `DlgTrackInfo`. This should probably be removed or reimplemented to
     /// check if all neighboring beats in this object have the same distance.
-    virtual bool hasConstantTempo() const = 0;
+    bool hasConstantTempo() const {
+        return m_markers.empty();
+    }
 
     /// Serialize beats to QByteArray.
-    virtual QByteArray toByteArray() const = 0;
+    QByteArray toByteArray() const;
 
     /// A string representing the version of the beat-processing code that
     /// produced this Beats instance. Used by BeatsFactory for associating a
     /// given serialization with the version that produced it.
-    virtual QString getVersion() const = 0;
+    QString getVersion() const;
     /// A sub-version can be used to represent the preferences used to generate
     /// the beats object.
-    virtual QString getSubVersion() const = 0;
+    QString getSubVersion() const {
+        return m_subVersion;
+    }
 
     ////////////////////////////////////////////////////////////////////////////
     // Beat calculations
@@ -111,10 +290,10 @@ class Beats : private std::enable_shared_from_this<Beats> {
     /// `position` refers to the location of the beat, the first value is
     /// `position`, and the second value is the next beat position. Returns
     /// `false` if *at least one* position is invalid.
-    virtual bool findPrevNextBeats(audio::FramePos position,
+    bool findPrevNextBeats(audio::FramePos position,
             audio::FramePos* prevBeatPosition,
             audio::FramePos* nextBeatPosition,
-            bool snapToNearBeats) const = 0;
+            bool snapToNearBeats) const;
 
     /// Return the frame position of the first beat in the track, or an invalid
     /// position if none exists.
@@ -133,7 +312,7 @@ class Beats : private std::enable_shared_from_this<Beats> {
     /// `findNextBeat` and `findPrevBeat`, respectively. If `position` refers
     /// to the location of a beat, then `position` is returned. If no beat can
     /// be found, returns an invalid frame position.
-    virtual audio::FramePos findNthBeat(audio::FramePos position, int n) const = 0;
+    audio::FramePos findNthBeat(audio::FramePos position, int n) const;
 
     int numBeatsInRange(audio::FramePos startPosition, audio::FramePos endPosition) const;
 
@@ -146,25 +325,27 @@ class Beats : private std::enable_shared_from_this<Beats> {
     /// while holding a strong references to the `Beats` object to ensure that
     /// the `Beats` object is not deleted. Caller takes ownership of the returned
     /// `BeatIterator`.
-    virtual std::unique_ptr<BeatIterator> findBeats(
+    std::unique_ptr<BeatIterator> findBeats(
             audio::FramePos startPosition,
-            audio::FramePos endPosition) const = 0;
+            audio::FramePos endPosition) const;
 
     /// Return whether or not a beat exists between `startPosition` and `endPosition`.
-    virtual bool hasBeatInRange(audio::FramePos startPosition,
-            audio::FramePos endPosition) const = 0;
+    bool hasBeatInRange(audio::FramePos startPosition,
+            audio::FramePos endPosition) const;
 
     /// Return the predominant BPM value between `startPosition` and `endPosition`
     /// if the BPM is valid, otherwise returns an invalid BPM value.
-    virtual mixxx::Bpm getBpmInRange(audio::FramePos startPosition,
-            audio::FramePos endPosition) const = 0;
+    mixxx::Bpm getBpmInRange(audio::FramePos startPosition,
+            audio::FramePos endPosition) const;
 
     /// Return the arithmetic average BPM over the range of n*2 beats centered around
     /// frame position `position`. For example, n=4 results in an averaging of 8 beats.
     /// The returned Bpm value may be invalid.
-    virtual mixxx::Bpm getBpmAroundPosition(audio::FramePos position, int n) const = 0;
+    mixxx::Bpm getBpmAroundPosition(audio::FramePos position, int n) const;
 
-    virtual audio::SampleRate getSampleRate() const = 0;
+    audio::SampleRate getSampleRate() const {
+        return m_sampleRate;
+    }
 
     ////////////////////////////////////////////////////////////////////////////
     // Beat mutations
@@ -176,19 +357,19 @@ class Beats : private std::enable_shared_from_this<Beats> {
     //
     /// Returns a pointer to the modified beats object, or `nullopt` on
     /// failure.
-    virtual std::optional<BeatsPointer> tryTranslate(audio::FrameDiff_t offset) const = 0;
+    std::optional<BeatsPointer> tryTranslate(audio::FrameDiff_t offset) const;
 
     /// Scale the position of every beat in the song by `scale`.
     //
     /// Returns a pointer to the modified beats object, or `nullopt` on
     /// failure.
-    virtual std::optional<BeatsPointer> tryScale(BpmScale scale) const = 0;
+    std::optional<BeatsPointer> tryScale(BpmScale scale) const;
 
     /// Adjust the beats so the global average BPM matches `bpm`.
     //
     /// Returns a pointer to the modified beats object, or `nullopt` on
     /// failure.
-    virtual std::optional<BeatsPointer> trySetBpm(mixxx::Bpm bpm) const = 0;
+    std::optional<BeatsPointer> trySetBpm(mixxx::Bpm bpm) const;
 
   protected:
     /// Type tag for making public constructors of derived classes inaccessible.
@@ -198,11 +379,41 @@ class Beats : private std::enable_shared_from_this<Beats> {
 
     Beats() = default;
 
-    virtual bool isValid() const = 0;
+    bool isValid() const;
 
   private:
     Beats(const Beats&) = delete;
     Beats(Beats&&) = delete;
+
+    QByteArray toBeatGridByteArray() const;
+    QByteArray toBeatMapByteArray() const;
+
+    mixxx::audio::FrameDiff_t firstBeatLengthFrames() const;
+    mixxx::audio::FrameDiff_t endBeatLengthFrames() const;
+
+    std::vector<BeatMarker> m_markers;
+    mixxx::audio::FramePos m_endMarkerPosition;
+    mixxx::Bpm m_endMarkerBpm;
+    mixxx::audio::SampleRate m_sampleRate;
+
+    // The sub-version of this beatgrid.
+    const QString m_subVersion;
+};
+
+class BeatIterator {
+  public:
+    BeatIterator(Beats::ConstIterator it, audio::FramePos endPosition)
+            : m_it(std::move(it)),
+              m_endPosition(endPosition) {
+    }
+
+    ~BeatIterator() = default;
+    bool hasNext() const;
+    audio::FramePos next();
+
+  private:
+    Beats::ConstIterator m_it;
+    const audio::FramePos m_endPosition;
 };
 
 } // namespace mixxx
